# routes/usersRoute.js
const router = require("express").Router();
const User = require("../models/userModel");
const bcrypt = require("bcryptjs");
const jwt = require("jsonwebtoken");
const authMiddleware = require("../middlewares/authMiddleware");

// user registration

router.post("/register", async (req, res) => {
  try {
    // check if user already exists
    const userExists = await User.findOne({ email: req.body.email });
    if (userExists) {
      return res
        .status(200)
        .send({ message: "User already exists", success: false });
    }

    // hash password
    const salt = await bcrypt.genSalt(10);
    const hashedPassword = await bcrypt.hash(req.body.password, salt);
    req.body.password = hashedPassword;

    // create new user
    const newUser = new User(req.body);
    await newUser.save();
    res.send({
      message: "User created successfully",
      success: true,
    });
  } catch (error) {
    res.status(500).send({
      message: error.message,
      data: error,
      success: false,
    });
  }
});

// user login

router.post("/login", async (req, res) => {
  try {
    // check if user exists
    const user = await User.findOne({ email: req.body.email });
    if (!user) {
      return res
        .status(200)
        .send({ message: "User does not exist", success: false });
    }

    // check password
    const validPassword = await bcrypt.compare(
      req.body.password,
      user.password
    );
    if (!validPassword) {
      return res
        .status(200)
        .send({ message: "Invalid password", success: false });
    }

    const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET, {
      expiresIn: "1d",
    });

    res.send({
      message: "User logged in successfully",
      success: true,
      data: token,
    });
  } catch (error) {
    res.status(500).send({
      message: error.message,
      data: error,
      success: false,
    });
  }
});

// get user info

router.post("/get-user-info", authMiddleware, async (req, res) => {
  try {
    if (req.user.id === "admin") {
      return res.send({
        message: "Admin info fetched successfully",
        success: true,
        data: {
          _id: "admin",
          name: "Admin",
          email: "admin@cui.com",
          isAdmin: true,
        },
      });
    }
    // Normal user
    const user = await User.findById(req.user.userId);
    if (!user) {
      return res.status(404).send({
        message: "User not found",
        success: false,
        data: null,
      });
    }
    res.send({
      message: "User info fetched successfully",
      success: true,
      data: user,
    });
  } catch (error) {
    res.status(500).send({
      message: error.message,
      data: error,
      success: false,
    });
  }
});

// Admin Login Route
router.post("/admin-login", async (req, res) => {
  try {
    const { email, password } = req.body;
    if (email !== "admin@cui.com" || password !== "admin") {
      return res.status(200).json({
        message: "Invalid admin credentials",
        success: false,
      });
    }
    const token = jwt.sign({ id: "admin" }, process.env.jwt_secret, {
      expiresIn: "1d",
    });
    res.status(200).json({
      message: "Admin login successful",
      success: true,
      data: token,
    });
  } catch (error) {
    res.status(500).json({
      message: error.message,
      success: false,
    });
  }
});

module.exports = router;